main.py

+++ main.py
@@ -1,5 +1,5 @@
-a=int(input("Enter First No:"))
-b=int(input("Enter Second No:"))
+a=float(input("Enter First No:"))
+b=float(input("Enter Second No:"))
 c=a+b
 print("addition Is: ",c)
